ArduCopter: move Auxiliary functions to AP_AuxFunc
