+++ ArduCopter/RC_Channel.cpp
@@ -1,0 +1,130 @@
+#include "Copter.h"
+
+#include "RC_Channel.h"
+
+
+// defining these two macros and including the RC_Channels_VarInfo header defines the parameter information common to all vehicle types
+#define RC_CHANNELS_SUBCLASS RC_Channels_Copter
+#define RC_CHANNEL_SUBCLASS RC_Channel_Copter
+
+#include <RC_Channel/RC_Channels_VarInfo.h>
+
+int8_t RC_Channels_Copter::flight_mode_channel_number() const
+{
+    return copter.g.flight_mode_chan.get();
+}
+
+void RC_Channel_Copter::mode_switch_changed(modeswitch_pos_t new_pos)
+{
+    if (new_pos < 0 || (uint8_t)new_pos > copter.num_flight_modes) {
+        // should not have been called
+        return;
+    }
+
+    if (!copter.set_mode((Mode::Number)copter.flight_modes[new_pos].get(), ModeReason::RC_COMMAND)) {
+        // alert user to mode change failure
+        if (copter.ap.initialised) {
+            AP_Notify::events.user_mode_change_failed = 1;
+        }
+        return;
+    }
+
+    // play a tone
+    // alert user to mode change (except if autopilot is just starting up)
+    if (copter.ap.initialised) {
+        AP_Notify::events.user_mode_change = 1;
+    }
+
+    if (!rc().find_channel_for_option(AP_AuxFunc::Function::SIMPLE_MODE) &&
+        !rc().find_channel_for_option(AP_AuxFunc::Function::SUPERSIMPLE_MODE)) {
+        // if none of the Aux Switches are set to Simple or Super Simple Mode then
+        // set Simple Mode using stored parameters from EEPROM
+        if (BIT_IS_SET(copter.g.super_simple, new_pos)) {
+            copter.set_simple_mode(Copter::SimpleMode::SUPERSIMPLE);
+        } else {
+            copter.set_simple_mode(BIT_IS_SET(copter.g.simple_modes, new_pos) ? Copter::SimpleMode::SIMPLE : Copter::SimpleMode::NONE);
+        }
+    }
+}
+
+bool RC_Channels_Copter::has_valid_input() const
+{
+    if (copter.failsafe.radio) {
+        return false;
+    }
+    if (copter.failsafe.radio_counter != 0) {
+        return false;
+    }
+    return true;
+}
+
+RC_Channel * RC_Channels_Copter::get_arming_channel(void) const
+{
+    return copter.channel_yaw;
+}
+
+// save_trim - adds roll and pitch trims from the radio to ahrs
+void Copter::save_trim()
+{
+    // save roll and pitch trim
+    float roll_trim = ToRad((float)channel_roll->get_control_in()/100.0f);
+    float pitch_trim = ToRad((float)channel_pitch->get_control_in()/100.0f);
+    ahrs.add_trim(roll_trim, pitch_trim);
+    AP::logger().Write_Event(LogEvent::SAVE_TRIM);
+    gcs().send_text(MAV_SEVERITY_INFO, "Trim saved");
+}
+
+// auto_trim - slightly adjusts the ahrs.roll_trim and ahrs.pitch_trim towards the current stick positions
+// meant to be called continuously while the pilot attempts to keep the copter level
+void Copter::auto_trim_cancel()
+{
+    auto_trim_counter = 0;
+    AP_Notify::flags.save_trim = false;
+    gcs().send_text(MAV_SEVERITY_INFO, "AutoTrim cancelled");
+}
+
+void Copter::auto_trim()
+{
+    if (auto_trim_counter > 0) {
+        if (copter.flightmode != &copter.mode_stabilize ||
+            !copter.motors->armed()) {
+            auto_trim_cancel();
+            return;
+        }
+
+        // flash the leds
+        AP_Notify::flags.save_trim = true;
+
+        if (!auto_trim_started) {
+            if (ap.land_complete) {
+                // haven't taken off yet
+                return;
+            }
+            auto_trim_started = true;
+        }
+
+        if (ap.land_complete) {
+            // landed again.
+            auto_trim_cancel();
+            return;
+        }
+
+        auto_trim_counter--;
+
+        // calculate roll trim adjustment
+        float roll_trim_adjustment = ToRad((float)channel_roll->get_control_in() / 4000.0f);
+
+        // calculate pitch trim adjustment
+        float pitch_trim_adjustment = ToRad((float)channel_pitch->get_control_in() / 4000.0f);
+
+        // add trim to ahrs object
+        // save to eeprom on last iteration
+        ahrs.add_trim(roll_trim_adjustment, pitch_trim_adjustment, (auto_trim_counter == 0));
+
+        // on last iteration restore leds and accel gains to normal
+        if (auto_trim_counter == 0) {
+            AP_Notify::flags.save_trim = false;
+            gcs().send_text(MAV_SEVERITY_INFO, "AutoTrim: Trims saved");
+        }
+    }
+}
