+++ ArduCopter/RC_Channel.h
@@ -1,8 +1,6 @@
 #pragma once
 
 #include <RC_Channel/RC_Channel.h>
-#include <AP_Motors/AP_Motors.h>
-#include "mode.h"
 
 class RC_Channel_Copter : public RC_Channel
 {
@@ -11,15 +9,7 @@ public:
 
 protected:
 
-    void init_aux_function(aux_func_t ch_option, AuxSwitchPos) override;
-    bool do_aux_function(aux_func_t ch_option, AuxSwitchPos) override;
-
 private:
-
-    void do_aux_function_armdisarm(const AuxSwitchPos ch_flag) override;
-    void do_aux_function_change_mode(const Mode::Number mode,
-                                     const AuxSwitchPos ch_flag);
-    void do_aux_function_change_air_mode(const AuxSwitchPos ch_flag);
 
     // called when the mode switch changes position:
     void mode_switch_changed(modeswitch_pos_t new_pos) override;
