+++ ArduCopter/heli.cpp
@@ -142,7 +142,7 @@ void Copter::heli_update_landing_swash()
 // returns zero if motor interlock auxiliary switch hasn't been defined
 float Copter::get_pilot_desired_rotor_speed() const
 {
-    RC_Channel *rc_ptr = rc().find_channel_for_option(RC_Channel::AUX_FUNC::MOTOR_INTERLOCK);
+    RC_Channel *rc_ptr = rc().find_channel_for_option(AP_AuxFunc::Function::MOTOR_INTERLOCK);
     if (rc_ptr != nullptr) {
         rc_ptr->set_range(1000);
         return (float)rc_ptr->get_control_in() * 0.001f;
